CPP06 제출 전 점검

+++ CPP06/ex00/Converter.hpp
@@ -6,7 +6,7 @@
 #include <cmath>
 
 class Converter{
-	double init;
+	long double init;
 	public:
 		Converter();
 		Converter(const Converter& c);
